pyproject.toml
[tool.poetry]
name = "authorization"
version = "0.1.0"
description = ""
authors = ["Daniil Kiselev <kiselyovda@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = {extras = ["email"], version = "^2.6.4"}
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic-settings = "^2.2.1"
pyjwt = "^2.8.0"
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
mypy = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
